pyproject.toml
[project]
name = "prediction-market-data"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "brokenaxes>=0.6.2",
    "cryptography>=46.0.3",
    "duckdb>=1.4.2",
    "httpx>=0.28.1",
    "imageio>=2.36.0",
    "kalshi-python>=2.1.4",
    "matplotlib>=3.9.4",
    "pandas>=2.3.3",
    "polymarket-py>=0.1.0",
    "py-clob-client>=0.18.0",
    "pyarrow>=18.0.0",
    "scipy>=1.13.1",
    "squarify>=0.4.4",
    "tenacity>=8.0.0",
    "tqdm>=4.67.1",
    "web3>=6.0.0",
    "simple-term-menu>=1.6.0",
    "python-dotenv>=1.2.1",
    "flask>=3.0.0",
    "pmxt>=2.22.1",
    "anthropic>=0.40.0",
    "tavily-python>=0.7.0",
    "websocket-client>=1.9.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.9.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
markers = ["slow: marks tests as slow (deselect with '-m \"not slow\"')"]
addopts = "-p test_runtime_plugin -p no:cacheprovider"
filterwarnings = [
    "ignore:websockets\\.legacy is deprecated:DeprecationWarning",
    "ignore:Precision loss occurred in moment calculation due to catastrophic cancellation\\.:RuntimeWarning",
    "ignore:Animation was deleted without rendering anything\\.:UserWarning",
]

[tool.ruff]
target-version = "py39"
line-length = 120
exclude = [".venv"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["src"]
